Setup nodemailer
Setup nodemailer. Got a nice welcome email template from a ChatGPT.

Next step: Set logic for sending OTPs

I tried to use SMTPs like Brevo, turboSMTP, MailSend, MailJet, SendPulse, etc
but faced issues at sign up with almost all of them. turboSMTP was the only one
that worked properly but it was using my personal email for sending emaails and
when I tried creating a new account there were a lot of errors. In the end,
I resorted to Gmail which is seriously discouraged but I don't really have a choice
right now because for these SMTPs to work properly, I need to have a domain registered
so maybe I'll use them in the future but for now Gmail ought to be enough