Deployed to Render successfully
There were ALOT of issues here but the major one was with the cookie. It was being rejected
because no "partitioned" attribute was set. It's fixed now, though I wonder why it was
causing so many problems when it isn't even widely supported. Why was I forced to add it
when it might cause problems with compatibility?